Add static to inline functions
This is needed to avoid a link error where the inline functions appear
missing at link time.
From c99 standard inline function should either be declared static or
have an extern instance in a c file for linking.
This fix is necessary to build with gcc 7; for some reason it was not
trigerred before.

Signed-off-by: Julien Viard de Galbert <julien@vdg.name>